+++ frontend/src/metabase/visualizations/components/Table.css
