+++ frontend/src/metabase/visualizations/components/TableInteractive.css
