+++ frontend/src/metabase/visualizations/components/TableInteractive.jsx
