+++ frontend/src/metabase/visualizations/components/TableSimple.jsx
