+++ frontend/src/metabase/visualizations/visualizations/AreaChart.jsx
@@ -2,8 +2,8 @@
 
 import React, { Component, PropTypes } from "react";
 
-import LineAreaBarChart from "./components/LineAreaBarChart.jsx";
-import { areaRenderer } from "./lib/LineAreaBarRenderer";
+import LineAreaBarChart from "../components/LineAreaBarChart.jsx";
+import { areaRenderer } from "../lib/LineAreaBarRenderer";
 
 export default class AreaChart extends LineAreaBarChart {
     static uiName = "Area";
