+++ frontend/src/metabase/visualizations/visualizations/BarChart.jsx
@@ -2,8 +2,8 @@
 
 import React, { Component, PropTypes } from "react";
 
-import LineAreaBarChart from "./components/LineAreaBarChart.jsx";
-import { barRenderer } from "./lib/LineAreaBarRenderer";
+import LineAreaBarChart from "../components/LineAreaBarChart.jsx";
+import { barRenderer } from "../lib/LineAreaBarRenderer";
 
 export default class BarChart extends LineAreaBarChart {
     static uiName = "Bar";
