+++ frontend/src/metabase/visualizations/visualizations/Funnel.jsx
