+++ frontend/src/metabase/visualizations/visualizations/LineChart.jsx
@@ -2,8 +2,8 @@
 
 import React, { Component, PropTypes } from "react";
 
-import LineAreaBarChart from "./components/LineAreaBarChart.jsx";
-import { lineRenderer } from "./lib/LineAreaBarRenderer";
+import LineAreaBarChart from "../components/LineAreaBarChart.jsx";
+import { lineRenderer } from "../lib/LineAreaBarRenderer";
 
 export default class LineChart extends LineAreaBarChart {
     static uiName = "Line";
