+++ frontend/src/metabase/visualizations/visualizations/Progress.jsx
