+++ frontend/src/metabase/visualizations/visualizations/RowChart.jsx
@@ -2,8 +2,8 @@
 
 import React, { Component, PropTypes } from "react";
 
-import LineAreaBarChart from "./components/LineAreaBarChart.jsx";
-import { rowRenderer } from "./lib/LineAreaBarRenderer";
+import LineAreaBarChart from "../components/LineAreaBarChart.jsx";
+import { rowRenderer } from "../lib/LineAreaBarRenderer";
 
 export default class BarChart extends LineAreaBarChart {
     static uiName = "Row Chart";
