+++ frontend/src/metabase/visualizations/visualizations/Scalar.css
