+++ frontend/src/metabase/visualizations/visualizations/Scalar.jsx
